Function tdameritradeclient::auth::getcodeweblink

pub fn getcodeweblink(clientid: &str, redirecturi: &str) -> String

used to get code manually from tdameritrade website with redirect URI as localhost as per the directions on developer.tdameritrade.com -> you must register an app to get clientid and redirecturi code can be used with gettokenfromcode to initiate a refreshtoken and token