Struct tdameritradeclient::auth::TDauth

pub struct TDauth { /* fields omitted */ }

These are tools to help manage authorization tokens with TD Ameritrade API

getcodeweblink is a link created from parameters registered with deverloper.tdameritrade.com.
The link can be used to return a code to use to request a valid token and refresh token. You will need to log in with your TDAmeritrade Credentials. If you use a redirect_uri that points to localhost than you will see the code returned in the query bar of the browser

new_fromcode will allow you to update tokens from the code retrieved in 1)

new_fromrefresh will allow you to update tokens from the refresh_token. The refresh_token will stay active for 90 days so you can save for reuse.

Implementations

impl TDauth

pub fn new_from_refresh(
refresh: &str,
clientid: &str,
refreshupdate: bool
) -> TDauth

create new TDauth with refresh_token and clientid if successful TDauth will carry new valid token if refreshupdate is true than refresh_token will also be updated

pub fn new_from_code(
code: &str,
clientid: &str,
redirecturi: &str,
codedecode: bool
) -> TDauth

create new TDauth with code, redirecturi and clientid if successful TDauth will carry both new refresh_token and new valid token

you can use decode=true if you did NOT decode it only useful if you are using the browser to get code from query string

pub fn resolve_token_from_refresh(&mut self, refreshupdate: bool) -> String

get /oauth2/token token endpoint returns an access token along with an refresh token using refresh_token grant type and retrieves new refresh_token (optional)

returns full response and updates TDauth struct

pub fn resolve_token_fromcode(&mut self, code: &str, codedecode: bool) -> String

get /oauth2/token token endpoint returns an access token along with an optional refresh token using authorization_code grant type and retrieves new refresh_token (response returned) while storing valid token inside client

returns full json response as text and updates TDAuth

if grabbing code from browser as per the instructions on developer.tdameritrade.com then you will need to decode it. As the code is encoded when put in post body.

you can use decode=true if you did NOT decode it only useful if you are using the browser to get code from query string

pub fn get_tokens(&self) -> (&str, &str)