Module tdameritradeclient::param

holds all the available query parameters used with the endpoints

Structs

Empty

No Parameters

Enums

Account

Query Parameters for /accounts/

History

Query Parameters for /marketdata/{symbol}/pricehistory

Instruments

Query Parameters for /instruments

MarketHours

Query Parameters for /marketdata/{market}/hours

Movers

Query Parameters for /marketdata/{index}/movers

OptionChain

Query Parameters for /marketdata/chains

Order

Query Parameters for /orders/

Quotes

Query Parameters for /marketdata/quotes

Transactions

Query Parameters for /account/transactions

Traits

Pair

Query parameters must implement Pair trait to allow us to use Url crate in underlying http client to create the proper query_string

Functions

convert_to_pairs

function to convert a collection of query parameters in to a vector of pairs