Enum tdameritradeclient::param::History

pub enum History<'a> {
    PeriodType(&'a str),
    Period(u8),
    FrequencyType(&'a str),
    Frequency(u8),
    StartDate(u64),
    EndDate(u64),
    NeedExendedHoursData(bool),
}

Query Parameters for /marketdata/{symbol}/pricehistory

Variants

PeriodType(&'a str)

The type of period to show. Valid values are day, month, year, or ytd (year to date). Default is day.

Period(u8)

Number of periods to show. Valid by PeriodType (* is default)

day: 1, 2, 3, 4, 5, 10*

month: 1*, 2, 3, 6

year: 1*, 2, 3, 5, 10, 15, 20

ytd: 1*

FrequencyType(&'a str)

the type of frequency with which a new candle is formed. Valid frequencyTypes by periodType (defaults marked with an asterisk):

day: minute*

month: daily, weekly*

year: daily, weekly, monthly*

ytd: daily, weekly*

Frequency(u8)

the number of the frequencyType to be included in each candle. Valid frequencies by frequencyType (defaults marked with an asterisk):

minute: 1*, 5, 10, 15, 30

daily: 1*

weekly: 1*

monthly: 1*

StartDate(u64)

Start date as milliseconds since epoch. If startDate and endDate are provided, period should not be provided.

EndDate(u64)

End date as milliseconds since epoch. If startDate and endDate are provided, period should not be provided. Default is previous trading day.

NeedExendedHoursData(bool)

true to return extended hours data, false for regular market hours only. Default is true