Trait tdameritradeclient::param::Pair

pub trait Pair<'a> {
    fn pair(self) -> (&'a str, String);
}

Query parameters must implement Pair trait to allow us to use Url crate in underlying http client to create the proper query_string

Required methods

fn pair(self) -> (&'a str, String)