Enum tdameritradeclient::request::Endpoint

pub enum Endpoint<'a> {

    UserPrincipals,
    Accounts,
    Account(&'a str),
    Order((&'a str, &'a str)),
    Orders(&'a str),
    SavedOrder((&'a str, &'a str)),
    SavedOrders(&'a str),
    Quotes,
    MarketHours(&'a str),
    Instruments,
    Instrument(&'a str),
    History(&'a str),
    OptionChain,
    Transactions(&'a str),
    Transaction((&'a str, &'a str)),
    Watchlists(&'a str),
    Watchlist((&'a str, &'a str)),
    Movers(&'a str),

}

specifies Endpoints for TD Ameritrade’s API

Variants

UserPrincipals

/userprincipals

get user information including all linked accounts

Accounts

/accounts

get all account information like ids, access level, market approval, etc.

Account(&'a str)

/accounts/{ACCOUNTID}

get specific account info

optional param::Account to include positions and orders

Order((&'a str, &'a str))

/accounts/{ACCOUNTID}/orders/{ORDERID}

get order details for specified account id for specific order id

delete to cancel order

put to replace order; need to include body that will replace original order

First parameter is ACCOUNTID and second parameter is ORDERID

Orders(&'a str)

/accounts/{ACCOUNTID}/orders

get order details for specified account

post body with order details to submit order

SavedOrder((&'a str, &'a str))

/accounts/{ACCOUNTID}/savedorders/{SAVEDORDERID}

get order details for specified account id for specific order id

delete to cancel order

put to replace order; need to include body that will replace original order

First parameter is ACCOUNTID and second parameter is SAVEDORDERID

SavedOrders(&'a str)

/accounts/{ACCOUNTID}/savedorders

get order details for specified account

post body with order details to submit order

Quotes

/marketdata/quotes

get quotes for a list of symbols. example: “SPY,IWM,QQQ”

MarketHours(&'a str)

/marketdata/{MARKET}/hours

get todays market hours for given market

MARKET can be EQUITY, OPTION, FUTURE, BOND, or FOREX

optional param::MarketHours to specify market hours for a specific date

Instruments

Search Instruments

use param::Instruments for supplying Search parameters

Instrument(&'a str)

Get Instrument by CUSIP

History(&'a str)

/marketdata/{SYMBOL}/pricehistory

search for instruments

additional query parameters need to be added from param::History Enum

OptionChain

/marketdata/chains

additional query parameters need to be added from param::OptionChain Enum

Transactions(&'a str)

/accounts/{ACCOUNTID}/transactions

get transactions in a specified Account

additional query parameters can be added from param::Transactions Enum

Transaction((&'a str, &'a str))

/accounts/{ACCOUNTID}/transactions/{TRANSACTIONID}

get in a specified Account a specified transaction by Id

First parameter is ACCOUNTID and second parameter is TRANSACTIONID

Watchlists(&'a str)

/accounts/{ACCOUNTID}/watchlists

get all watchlists for specified account

post to create watchlist

Watchlist((&'a str, &'a str))

/accounts/{ACCOUNTID}/watchlists/{WATCHLISTID}

get a specific watchlid for a specified account

put to replace watchlist

patch to update watchlist

delete to remove watchlist

Movers(&'a str)

Has not been tested

/marketdata/{INDEX}/movers

get mover information by index symbol, direction type and change

INDEX: $SPX.X, $COMPX, $DJI

Implementations

impl<'a> Endpoint<'a>

pub fn url_endpoint(&self) -> String

defines the URL for the specified Endpoint