Struct tdameritradeclient::TDAClient

pub struct TDAClient { /* fields omitted */ }

Main client to access TD Ameritrade endpoints

Uses attohttpc::RequestBuilder to build requests and attohttpc::Session to maintain the same client configuration

Two options for output:

text which in this case is JSON from TDA API

convert to serde_json::Value

Implementations

impl TDAClient

pub fn new(token: String) -> TDAClient

Create new base client that maintains Authorization Header Requires valid token from tdameritrade

pub fn new_usingrefresh(refresh: &str, clientid: &str) -> TDAClient

Create new base client that maintains Authorization Header Requires valid refresh token from tdameritrade

pub fn new_usingcode(
code: &str,
clientid: &str,
redirecturi: &str,
codedecode: bool
) -> TDAClient

Create new base client that maintains Authorization Header Requires valid auth code from tdameritrade

you can use decode=true if you did NOT decode it only useful if you are using the browser to get code from query string

pub fn session_timeout(&mut self, duration: Duration)

change timeout configuration of Session

pub fn get_user_principals<T>(&self) -> T where
RequestBuilder: Execute<T>,

get /userprincipals

pub fn get_quotes<T>(&self, quotequery: &str) -> T where
RequestBuilder: Execute<T>,

get /marketdata/quotes?symbol=SYM1,SYM2,SYM3....

pub fn get_todays_market_hours<T>(&self, market: &str) -> T where
RequestBuilder: Execute<T>,

get /marketdata/{MARKET}/hours retrieve todays market hours for given market

pub fn get_dates_market_hours<T>(&self, market: &str, date: &str) -> T where
RequestBuilder: Execute<T>,

get /marketdata/{MARKET}/hours retrieve market hours for given market and date

pub fn get_instruments<T>(&self, params: &[Instruments<'_>]) -> T where
RequestBuilder: Execute<T>,

get /instruments

Search or retrieve instrument data, including fundamental data.

pub fn get_instrument<T>(&self, cusip: &str) -> T where
RequestBuilder: Execute<T>,

get /instruments/cusip

Get an instrument by CUSIP

pub fn get_history<T>(&self, symbol: &str, params: &[History<'_>]) -> T where
RequestBuilder: Execute<T>,

get /marketdata/{SYM}/pricehistory additional query parameters need to be added from History Enum retrieved based on EPOCH datetime also History Enum starttime and endtime is in EPOCH

pub fn get_option_chain<T>(&self, params: &[OptionChain<'_>]) -> T where
RequestBuilder: Execute<T>,

get /marketdata/chains?symbol=SYM additional query parameters need to be added from OptionChain Enum

pub fn get_accounts<T>(&self) -> T where
RequestBuilder: Execute<T>,

get /accounts if there are more than one account linked than it will retrieve an array of accounts

pub fn get_account<T>(&self, account: &str, params: &[Account]) -> T where
RequestBuilder: Execute<T>,

get /accounts/{account} grabs one account with account_id additional query parameters need to be added from Account Enum

pub fn get_orders<T>(&self, account: &str, params: &[Order<'_>]) -> T where
RequestBuilder: Execute<T>,

get /accounts/{account}/orders retrieve all working orders

pub fn get_transactions<T>(
&self,
account: &str,
params: &[Transactions<'_>]
) -> T where
RequestBuilder: Execute<T>,

get /accounts/{account}/transactions retrieve a specified transaction by Id

pub fn get_transaction<T>(&self, account: &str, transaction: &str) -> T where
RequestBuilder: Execute<T>,

get /accounts/{account}/transactions/{transactionId} retrieve a specified transaction by Id

pub fn get_watchlists<T>(&self, account: &str) -> T where
RequestBuilder: Execute<T>,

get /accounts/{account}/watchlists retrieves all watchlists for an account

pub fn create_order(&self, account: &str, ordertxt: &str) -> String

Post /accounts/{account}/orders with JSON formated body Creates a working order if JSON body has error it will return json indicating what's wrong if nothing is returned than request was good - could add additional error checking for 201 or 200 response

pub fn delete_order<T>(&self, account: &str, order: &str) -> T where
RequestBuilder: Execute<T>,

Delete /accounts/{account}/orders/{order} Creates a working order

pub fn replace_order(
&self,
account: &str,
order: &str,
ordertxt: &str
) -> String

PUT /accounts/{account}/orders/{order} with JSON formated body Replaces a working order with new order - allow the API to cancel and then creates new order