Module tdameritradeclient::auth

utility module to help with authorization token, refresh token and grant code

You can use the public functions or the TDauth struct. TDauth allows you to store the information for reuse.

Structs

TDauth
These are tools to help manage authorization tokens with TD Ameritrade API

Functions

get_code_weblink
used to get code manually from tdameritrade website with redirect URI as localhost as per the directions on developer.tdameritrade.com -> you must register an app to get clientid and redirecturi code can be used with gettokenfromcode to initiate a refreshtoken and token

get_refresh_from_refresh
used to get a valid refresh from refresh_token and clientid

get_token_from_code
used to get a valid token and refresh_token from code, clientid and redirecturi

get_token_from_refresh
used to get a valid token from refresh_token and clientid