Module tdameritradeclient::auth

Structs

TDauth
TDauth

Functions

getcodeweblink
used to get code manually from tdameritrade website with redirect URI as localhost as per the directions on developer.tdameritrade.com -> you must register an app to get clientid and redirecturi code can be used with gettokenfromcode to initiate a refreshtoken and token

getrefresh_fromrefresh
used to get a valid refresh from refresh_token and clientid

gettoken_fromcode
used to get a valid token and refresh_token from code, clientid and redirecturi

gettoken_fromrefresh
used to get a valid token from refresh_token and clientid